utils: split out source location helper from vlog
This allows oncore to use the constexpr feature for getting the filename
without having to include all of vlog formatting and logging bits.
